refactor(companies): move as component
