+++ src/pages/_home/_config.jsx
@@ -49,76 +49,3 @@ export const concepts = [
         description: translate({ id: "pages.home.concepts.needs.description" }),
     },
 ];
-
-/**
- * Companies using Feature-Sliced Design
- * @see https://github.com/feature-sliced/documentation/issues/131
- * @see static/img/companies
- */
-export const companies = [
-    {
-        url: "https://redmadrobot.com/",
-        src: "red_mad_robot.png",
-        alt: "red_mad_robot",
-    },
-    {
-        url: "https://kode.ru/",
-        src: "kode.png",
-        alt: "KODE",
-    },
-    {
-        url: "https://uptarget.co/",
-        src: "uptarget.png",
-        alt: "Uptarget",
-    },
-    {
-        url: "https://www.fxdd.com/",
-        src: "fxdd.svg",
-        alt: "FXDD",
-    },
-    {
-        url: "https://foxford.ru/",
-        src: "foxford.svg",
-        alt: "Foxford",
-    },
-    {
-        url: "https://vigo.ru/",
-        src: "vigo.png",
-        alt: "Vigo",
-    },
-    {
-        url: "https://space307.com/",
-        src: "space307.svg",
-        alt: "~/.space307",
-    },
-    {
-        url: "https://pmp-tech.ru/",
-        src: "pmp-tech.png",
-        alt: "PMP Tech",
-    },
-    {
-        url: "https://www.aligntech.com/",
-        src: "align.svg",
-        alt: "Align",
-    },
-    {
-        url: "https://smile2impress.com/",
-        src: "impress.svg",
-        alt: "Impress",
-    },
-    {
-        url: "https://pochta.tech/",
-        src: "pochtatech.svg",
-        alt: "Почтатех",
-    },
-    {
-        url: "https://befree.ru/",
-        src: "befree.svg",
-        alt: "befree",
-    },
-    {
-        url: "https://www.softcery.com/",
-        src: "softcery.png",
-        alt: "Softcery",
-    },
-];
