+++ src/pages/_home/styles.module.scss
@@ -13,48 +13,3 @@
         width: 100%;
     }
 }
-
-.companies {
-    --logoCardSize: 150px;
-    --logoCardGap: 3rem;
-
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(var(--logoCardSize), 1fr));
-    grid-gap: var(--logoCardGap);
-    width: 100%;
-    padding: 0 5% 40px 5%;
-
-    /* FIXME: fix later */
-    &Item {
-        align-self: center;
-        transition: 0.25s;
-
-        &:hover {
-            opacity: 0.7;
-        }
-    }
-
-    &ItemImg {
-        width: 100%;
-    }
-
-    &Subtitle {
-        width: 100%;
-        font-size: 12px;
-        font-style: italic;
-        text-align: center;
-    }
-
-    /* Protected from dark mode (and flickering) */
-    &Container {
-        color: #1c1e21;
-        background: #ffffff;
-    }
-}
-
-@media screen and (max-width: 966px) {
-    .companies {
-        --logoCardSize: 150px;
-        --logoCardGap: 1rem;
-    }
-}
